+++ experimental/packages/opentelemetry-instrumentation/src/instrumentation.ts
@@ -25,16 +25,19 @@ import {
   TracerProvider,
 } from '@opentelemetry/api';
 import * as shimmer from 'shimmer';
-import { InstrumentationModuleDefinition } from './platform/node';
-import * as types from './types';
+import {
+  InstrumentationModuleDefinition,
+  Instrumentation,
+  InstrumentationConfig,
+} from './types';
 
 /**
  * Base abstract internal class for instrumenting node and web plugins
  */
 export abstract class InstrumentationAbstract<T = any>
-  implements types.Instrumentation
+  implements Instrumentation
 {
-  protected _config: types.InstrumentationConfig;
+  protected _config: InstrumentationConfig;
 
   private _tracer: Tracer;
   private _meter: Meter;
@@ -43,7 +46,7 @@ export abstract class InstrumentationAbstract<T = any>
   constructor(
     public readonly instrumentationName: string,
     public readonly instrumentationVersion: string,
-    config: types.InstrumentationConfig = {}
+    config: InstrumentationConfig = {}
   ) {
     this._config = {
       enabled: true,
@@ -95,7 +98,7 @@ export abstract class InstrumentationAbstract<T = any>
   }
 
   /* Returns InstrumentationConfig */
-  public getConfig(): types.InstrumentationConfig {
+  public getConfig(): InstrumentationConfig {
     return this._config;
   }
 
@@ -103,7 +106,7 @@ export abstract class InstrumentationAbstract<T = any>
    * Sets InstrumentationConfig to this plugin
    * @param InstrumentationConfig
    */
-  public setConfig(config: types.InstrumentationConfig = {}): void {
+  public setConfig(config: InstrumentationConfig = {}): void {
     this._config = Object.assign({}, config);
   }
