+++ experimental/packages/opentelemetry-instrumentation/src/platform/browser/index.ts
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-export * from './instrumentation';
+export { InstrumentationBase } from './instrumentation';
