+++ experimental/packages/opentelemetry-instrumentation/src/platform/index.ts
@@ -14,10 +14,4 @@
  * limitations under the License.
  */
 
-export {
-  InstrumentationBase,
-  InstrumentationModuleDefinition,
-  InstrumentationModuleFile,
-  InstrumentationNodeModuleDefinition,
-  InstrumentationNodeModuleFile,
-} from './node';
+export { InstrumentationBase } from './node';
