+++ experimental/packages/opentelemetry-instrumentation/src/platform/node/index.ts
@@ -14,9 +14,3 @@
  * limitations under the License.
  */
 export { InstrumentationBase } from './instrumentation';
-export { InstrumentationNodeModuleDefinition } from './instrumentationNodeModuleDefinition';
-export { InstrumentationNodeModuleFile } from './instrumentationNodeModuleFile';
-export {
-  InstrumentationModuleDefinition,
-  InstrumentationModuleFile,
-} from './types';
