+++ experimental/packages/opentelemetry-instrumentation/src/platform/node/instrumentation.ts
@@ -26,7 +26,7 @@ import {
 } from './RequireInTheMiddleSingleton';
 import type { HookFn } from 'import-in-the-middle';
 import * as ImportInTheMiddle from 'import-in-the-middle';
-import { InstrumentationModuleDefinition } from './types';
+import { InstrumentationModuleDefinition } from '../../types';
 import { diag } from '@opentelemetry/api';
 import type { OnRequireFn } from 'require-in-the-middle';
 import { Hook } from 'require-in-the-middle';
